# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Start API",
        "description": "AgroBozor uchun backend API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.daladan.uz/api/v1/",
            "description": "Production server"
        },
        {
            "url": "http://daladan-api.loc/api/v1",
            "description": "Local dev server (OSPanel)"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "php artisan serve"
        }
    ],
    "paths": {
        "/admin/ad-promotions/{promotion}/confirm": {
            "patch": {
                "tags": [
                    "Admin Ads Moderation"
                ],
                "summary": "Pending promo buyurtmasini tasdiqlash (active + ads sinxron)",
                "description": "Click/Payme dan keyin ixtiyoriy payment_transaction_id. Faqat pending holat.",
                "operationId": "a64f01a395712e8faa8fc7bc32071769",
                "parameters": [
                    {
                        "name": "promotion",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminAdPromotionConfirmPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tasdiqlandi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminAdPromotionConfirmSuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Holat yoki validatsiya"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/ads": {
            "get": {
                "tags": [
                    "Admin Ads Moderation"
                ],
                "summary": "E'lonlar (moderatsiya ro'yxati)",
                "operationId": "04ab4c0a8689b92228d972b957753b45",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "pending|active|rejected|sold|deleted",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginatsiyali e'lonlar"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/ads/{id}": {
            "get": {
                "tags": [
                    "Admin Ads Moderation"
                ],
                "summary": "Bitta e'lon",
                "operationId": "e792fe7d9a47c0afe6270e0bd5d882c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "E'lon obyekti"
                    },
                    "404": {
                        "description": "Topilmadi"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/ads/{id}/edit": {
            "patch": {
                "tags": [
                    "Admin Ads Moderation"
                ],
                "summary": "Pending e'lonni tahrirlash",
                "operationId": "dbe7d9f37b42a0142581afd06bee84c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yangilangan e'lon"
                    },
                    "404": {
                        "description": "Topilmadi"
                    },
                    "422": {
                        "description": "Faqat pending yoki validatsiya"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/ads/{id}/approve": {
            "patch": {
                "tags": [
                    "Admin Ads Moderation"
                ],
                "summary": "Pending e'lonni tasdiqlash (active)",
                "operationId": "8536fb4ef276dba91492776ae8263015",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasdiq xabari"
                    },
                    "404": {
                        "description": "Topilmadi"
                    },
                    "422": {
                        "description": "Faqat pending"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/ads/{id}/reject": {
            "patch": {
                "tags": [
                    "Admin Ads Moderation"
                ],
                "summary": "Pending e'lonni rad etish",
                "operationId": "9413acf2f1dbe10615a6e62c08ca2ccb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "maxLength": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rad etildi"
                    },
                    "404": {
                        "description": "Topilmadi"
                    },
                    "422": {
                        "description": "Faqat pending yoki reason yo'q"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AdminAdPromotionConfirmPayload": {
                "properties": {
                    "payment_transaction_id": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "click_abc123",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AdminAdPromotionAdSnippet": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "title": {
                        "type": "string",
                        "example": "Naslli echkilar"
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
            "AdminAdPromotionPlanSnippet": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Top sotuv — 7 kun"
                    },
                    "slug": {
                        "type": "string",
                        "example": "top-sale-7d"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "top_sale",
                            "boost"
                        ]
                    }
                },
                "type": "object"
            },
            "AdminAdPromotionUserSnippet": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "phone": {
                        "type": "string",
                        "example": "+998901234567"
                    },
                    "fname": {
                        "type": "string",
                        "example": "Ali"
                    }
                },
                "type": "object"
            },
            "AdminAdPromotionDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "ad_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "promotion_plan_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "amount_paid": {
                        "type": "number",
                        "format": "float",
                        "example": 9000
                    },
                    "currency": {
                        "type": "string",
                        "example": "UZS"
                    },
                    "started_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "payment_transaction_id": {
                        "type": "string",
                        "nullable": true
                    },
                    "ad": {
                        "$ref": "#/components/schemas/AdminAdPromotionAdSnippet"
                    },
                    "promotion_plan": {
                        "$ref": "#/components/schemas/AdminAdPromotionPlanSnippet"
                    },
                    "user": {
                        "$ref": "#/components/schemas/AdminAdPromotionUserSnippet"
                    }
                },
                "type": "object"
            },
            "AdminAdPromotionConfirmSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AdminAdPromotionDetail"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}